ml-pipeline/pyproject.toml
[project]
name = "ml-pipeline"
version = "0.1.0"
description = "Succulent genus classification using EfficientNet V2 + ArcFace"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
]

[project.scripts]
create-splits = "create_splits:main"
train = "train:main"
evaluate = "evaluate:main"
